add feature order with midtrans,  notification, and fix some minor bug

+++ src/main/java/com/binar/byteacademy/entity/Promo.java
@@ -31,7 +31,7 @@ public class Promo {
     private String promoCode;
 
     @Column(nullable = false)
-    private Double discount;
+    private Integer discount;
 
     @Column(columnDefinition = "text")
     private String promoDescription;
